Extract OwnerReference handlers to own packages
Cluster & MachinePool controllers need to be configured with more handlers so
it makes sense to extract embedded handlers to their own packages as others.
