+++ service/controller/cluster.go
@@ -2,7 +2,6 @@ package controller
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/giantswarm/k8sclient/v3/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
@@ -13,14 +12,11 @@ import (
 	"github.com/giantswarm/operatorkit/resource/wrapper/retryresource"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
-	"sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
 	capiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	"github.com/giantswarm/azure-operator/v4/pkg/label"
 	"github.com/giantswarm/azure-operator/v4/pkg/project"
-	"github.com/giantswarm/azure-operator/v4/service/controller/key"
+	"github.com/giantswarm/azure-operator/v4/service/controller/resource/clusterownerreference"
 	"github.com/giantswarm/azure-operator/v4/service/controller/setting"
 )
 
@@ -90,13 +86,13 @@ func NewClusterResourceSet(config ClusterConfig) ([]resource.Interface, error) {
 
 	var ownerReferencesResource resource.Interface
 	{
-		c := ClusterOwnerReferencesConfig{
+		c := clusterownerreference.Config{
 			CtrlClient: config.K8sClient.CtrlClient(),
 			Logger:     config.Logger,
 			Scheme:     config.K8sClient.Scheme(),
 		}
 
-		ownerReferencesResource, err = NewClusterOwnerReferences(c)
+		ownerReferencesResource, err = clusterownerreference.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -126,82 +122,4 @@ func NewClusterResourceSet(config ClusterConfig) ([]resource.Interface, error) {
 	}
 
 	return resources, nil
-}
-
-type ClusterOwnerReferencesConfig struct {
-	CtrlClient client.Client
-	Logger     micrologger.Logger
-	Scheme     *runtime.Scheme
-}
-
-type ClusterOwnerReferencesResource struct {
-	ctrlClient client.Client
-	logger     micrologger.Logger
-	scheme     *runtime.Scheme
-}
-
-func NewClusterOwnerReferences(config ClusterOwnerReferencesConfig) (*ClusterOwnerReferencesResource, error) {
-	if config.CtrlClient == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
-	}
-	if config.Logger == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
-	}
-	if config.Scheme == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Scheme must not be empty", config)
-	}
-
-	r := &ClusterOwnerReferencesResource{
-		ctrlClient: config.CtrlClient,
-		logger:     config.Logger,
-		scheme:     config.Scheme,
-	}
-
-	return r, nil
-}
-
-// EnsureCreated ensures the AzureCluster is owned by the Cluster it belongs to.
-func (r *ClusterOwnerReferencesResource) EnsureCreated(ctx context.Context, obj interface{}) error {
-	cluster, err := key.ToCluster(obj)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	r.logger.LogCtx(ctx, "message", fmt.Sprintf("Ensuring %s label and 'ownerReference' fields on AzureCluster '%s/%s'", capiv1alpha3.ClusterLabelName, cluster.Namespace, cluster.Spec.InfrastructureRef.Name))
-
-	azureCluster := v1alpha3.AzureCluster{}
-	err = r.ctrlClient.Get(ctx, client.ObjectKey{Namespace: cluster.Namespace, Name: cluster.Spec.InfrastructureRef.Name}, &azureCluster)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	if azureCluster.Labels == nil {
-		azureCluster.Labels = make(map[string]string)
-	}
-	azureCluster.Labels[capiv1alpha3.ClusterLabelName] = cluster.Name
-
-	// Set Cluster as owner of AzureCluster
-	err = controllerutil.SetControllerReference(&cluster, &azureCluster, r.scheme)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	err = r.ctrlClient.Update(ctx, &azureCluster)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	r.logger.LogCtx(ctx, "message", fmt.Sprintf("Ensured %s label and 'ownerReference' fields on AzureCluster '%s/%s'", capiv1alpha3.ClusterLabelName, cluster.Namespace, cluster.Spec.InfrastructureRef.Name))
-
-	return nil
-}
-
-// EnsureDeleted is a noop.
-func (r *ClusterOwnerReferencesResource) EnsureDeleted(ctx context.Context, obj interface{}) error {
-	return nil
-}
-
-// Name returns the resource name.
-func (r *ClusterOwnerReferencesResource) Name() string {
-	return "ClusterOwnerReferences"
 }
