+++ service/controller/machine_pool.go
@@ -2,7 +2,6 @@ package controller
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/giantswarm/k8sclient/v3/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
@@ -11,19 +10,13 @@ import (
 	"github.com/giantswarm/operatorkit/resource"
 	"github.com/giantswarm/operatorkit/resource/wrapper/metricsresource"
 	"github.com/giantswarm/operatorkit/resource/wrapper/retryresource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
-	expcapzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
-	capiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
 	expcapiv1alpha3 "sigs.k8s.io/cluster-api/exp/api/v1alpha3"
-	capiutil "sigs.k8s.io/cluster-api/util"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	"github.com/giantswarm/azure-operator/v4/pkg/label"
 	"github.com/giantswarm/azure-operator/v4/pkg/project"
-	"github.com/giantswarm/azure-operator/v4/service/controller/key"
+	"github.com/giantswarm/azure-operator/v4/service/controller/resource/machinepoolownerreference"
 )
 
 type MachinePoolConfig struct {
@@ -92,13 +85,13 @@ func NewMachinePoolResourceSet(config MachinePoolConfig) ([]resource.Interface, 
 
 	var ownerReferencesResource resource.Interface
 	{
-		c := MachinePoolOwnerReferencesConfig{
+		c := machinepoolownerreference.Config{
 			CtrlClient: config.K8sClient.CtrlClient(),
 			Logger:     config.Logger,
 			Scheme:     config.K8sClient.Scheme(),
 		}
 
-		ownerReferencesResource, err = NewMachinePoolOwnerReferences(c)
+		ownerReferencesResource, err = machinepoolownerreference.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -128,108 +121,4 @@ func NewMachinePoolResourceSet(config MachinePoolConfig) ([]resource.Interface, 
 	}
 
 	return resources, nil
-}
-
-type MachinePoolOwnerReferencesConfig struct {
-	CtrlClient client.Client
-	Logger     micrologger.Logger
-	Scheme     *runtime.Scheme
-}
-
-type MachinePoolOwnerReferencesResource struct {
-	ctrlClient client.Client
-	logger     micrologger.Logger
-	scheme     *runtime.Scheme
-}
-
-func NewMachinePoolOwnerReferences(config MachinePoolOwnerReferencesConfig) (*MachinePoolOwnerReferencesResource, error) {
-	if config.CtrlClient == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
-	}
-	if config.Logger == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
-	}
-	if config.Scheme == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Scheme must not be empty", config)
-	}
-
-	r := &MachinePoolOwnerReferencesResource{
-		ctrlClient: config.CtrlClient,
-		logger:     config.Logger,
-		scheme:     config.Scheme,
-	}
-
-	return r, nil
-}
-
-// EnsureCreated ensures the MachinePool is owned by the Cluster it belongs to, and the AzureMachinePool is owned by the MachinePool.
-func (r *MachinePoolOwnerReferencesResource) EnsureCreated(ctx context.Context, obj interface{}) error {
-	machinePool, err := key.ToMachinePool(obj)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	r.logger.LogCtx(ctx, "message", fmt.Sprintf("Ensuring %s label and 'ownerReference' fields on MachinePool '%s/%s'", capiv1alpha3.ClusterLabelName, machinePool.Namespace, machinePool.Name))
-	r.logger.LogCtx(ctx, "message", fmt.Sprintf("Ensuring %s label and 'ownerReference' fields on AzureMachinePool '%s/%s'", capiv1alpha3.ClusterLabelName, machinePool.Namespace, machinePool.Spec.Template.Spec.InfrastructureRef.Name))
-
-	azureMachinePool := expcapzv1alpha3.AzureMachinePool{}
-	err = r.ctrlClient.Get(ctx, client.ObjectKey{Namespace: machinePool.Namespace, Name: machinePool.Spec.Template.Spec.InfrastructureRef.Name}, &azureMachinePool)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	if machinePool.Labels == nil {
-		machinePool.Labels = make(map[string]string)
-	}
-	machinePool.Labels[capiv1alpha3.ClusterLabelName] = machinePool.Spec.ClusterName
-
-	if azureMachinePool.Labels == nil {
-		azureMachinePool.Labels = make(map[string]string)
-	}
-	azureMachinePool.Labels[capiv1alpha3.ClusterLabelName] = machinePool.Spec.ClusterName
-
-	cluster, err := capiutil.GetClusterByName(ctx, r.ctrlClient, machinePool.Namespace, machinePool.Spec.ClusterName)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	// Set Cluster as owner of MachinePool
-	machinePool.OwnerReferences = capiutil.EnsureOwnerRef(machinePool.OwnerReferences, metav1.OwnerReference{
-		APIVersion: capiv1alpha3.GroupVersion.String(),
-		Kind:       "Cluster",
-		Name:       cluster.Name,
-		UID:        cluster.UID,
-	})
-
-	err = r.ctrlClient.Update(ctx, &machinePool)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	r.logger.LogCtx(ctx, "message", fmt.Sprintf("Ensured %s label and 'ownerReference' fields on MachinePool '%s/%s'", capiv1alpha3.ClusterLabelName, machinePool.Namespace, machinePool.Name))
-
-	// Set MachinePool as owner of AzureMachinePool
-	err = controllerutil.SetControllerReference(&machinePool, &azureMachinePool, r.scheme)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	err = r.ctrlClient.Update(ctx, &azureMachinePool)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	r.logger.LogCtx(ctx, "message", fmt.Sprintf("Ensuring %s label and 'ownerReference' fields on AzureMachinePool '%s/%s'", capiv1alpha3.ClusterLabelName, machinePool.Namespace, machinePool.Spec.Template.Spec.InfrastructureRef.Name))
-
-	return nil
-}
-
-// EnsureDeleted is a noop.
-func (r *MachinePoolOwnerReferencesResource) EnsureDeleted(ctx context.Context, obj interface{}) error {
-	return nil
-}
-
-// Name returns the resource name.
-func (r *MachinePoolOwnerReferencesResource) Name() string {
-	return "MachinePoolOwnerReferences"
 }
